refactor: re-ducks constructure
