+++ App.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Provider } from 'react-redux'
 import { AppearanceProvider, useColorScheme } from 'react-native-appearance'
-import configureStore from './src/reducers/configureStore'
+import configureStore from './src/store/configureStore'
 import AppNavigator from './src/navigators/AppNavigator'
 import { LoadingModal } from './src/components/organisms'
