+++ src/screens/HomeScreen.tsx
@@ -5,7 +5,7 @@ import { headerNavigationOptions } from '../navigators/options'
 import { useParties } from '../services/party'
 import { useModal } from '../services/modal'
 import { useStyles, MakeStyles } from '../services/design'
-import { useAuthState } from '../reducers'
+import { useAuthState } from '../store/hooks'
 import { LoadingPage } from '../components/pages'
 import { Card, GenderModal, Modal } from '../components/organisms'
 import { setGender } from '../services/user'
