+++ src/screens/PartyEntryScreen.tsx
@@ -5,7 +5,7 @@ import { User } from '../entities'
 import { View, Text, StyleSheet, Dimensions } from 'react-native'
 import { useModal } from '../services/modal'
 import { useStyles, MakeStyles } from '../services/design'
-import { useAuthState } from '../reducers'
+import { useAuthState } from '../store/hooks'
 import { entryPartyMembers } from '../services/party'
 import * as userRepository from '../repositories/user'
 import { LoadingPage, SearchUserPage } from '../components/pages'
