+++ src/screens/SettingScreen.tsx
@@ -1,8 +1,7 @@
 import React, { useCallback } from 'react'
 import { NavigationStackProp, NavigationStackScreenProps } from 'react-navigation-stack'
 import { headerNavigationOptions } from '../navigators/options'
-import { useAuthState } from '../reducers'
-import { useAuthActions } from '../actions'
+import { useAuthState, useAuthActions } from '../store/hooks'
 import { useStyles, MakeStyles } from '../services/design'
 import { useSecure, usePushNotifications } from '../services/secure'
 import { TouchableOpacity, ScrollView, Text, StyleSheet, Dimensions } from 'react-native'
