+++ src/screens/UserEditScreen.tsx
@@ -2,8 +2,7 @@ import React, { useCallback } from 'react'
 import { NavigationStackProp, NavigationStackScreenProps } from 'react-navigation-stack'
 import { headerNavigationOptions } from '../navigators/options'
 import { UpdateUser } from '../entities'
-import { useAuthState } from '../reducers'
-import { useUIActions } from '../actions'
+import { useAuthState, useUIActions } from '../store/hooks'
 import * as UserRepository from '../repositories/user'
 import { useStyles, useColors, MakeStyles } from '../services/design'
 import { useUserEditTools } from '../services/user'
