+++ src/store/configureSaga.ts
@@ -1,6 +1,6 @@
 import { all } from 'redux-saga/effects'
-import auth from './auth'
+import { authSaga } from './auth'
 
 export default function* rootSaga() {
-  yield all([...auth])
+  yield all([...authSaga])
 }
