+++ src/store/hooks.ts
@@ -1,0 +1,2 @@
+export { useAuthState, useAuthActions } from './auth'
+export { useUIState, useUIActions } from './ui'
